# api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/api.proto",
    "version": "0.0.0",
    "contact": {
      "name": "NoCloud",
      "url": "https://github.com/slntopp/nocloud",
      "email": "info@slnt-opp.xyz"
    }
  },
...
  "host": "localhost:8000",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json",
    "text/event-stream"
  ],
  "produces": [
    "application/json",
    "text/event-stream"
  ],
  "paths": {
...
    "/settings": {
      "get": {
        "operationId": "SettingsService_Keys",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/settingsKeysResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "SettingsService"
        ]
      },
      "post": {
        "operationId": "SettingsService_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/nocloudsettingsGetRequest"
            }
          }
        ],
        "tags": [
          "SettingsService"
        ]
      }
    },
    "/settings/watch": {
      "get": {
        "operationId": "SettingsService_Sub",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/settingsKeyEvent"
                },
                "error": {
                  "$ref": "#/definitions/googlerpcStatus"
                }
              },
              "title": "Stream result of settingsKeyEvent"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "keys",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "SettingsService"
        ]
      }
    },
    "/settings/{key}": {
      "delete": {
        "operationId": "SettingsService_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/nocloudsettingsDeleteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "SettingsService"
        ]
      },
      "put": {
        "operationId": "SettingsService_Put",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/settingsPutResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "public": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "tags": [
          "SettingsService"
        ]
      }
    },
...
  },
  "definitions": {
    "KeysResponseKey": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "public": {
          "type": "boolean"
        }
      }
    },
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "nocloudsettingsDeleteResponse": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        }
      }
    },
    "nocloudsettingsGetRequest": {
      "type": "object",
      "properties": {
        "keys": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "settingsKeyEvent": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "event": {
          "type": "string"
        }
      }
    },
    "settingsKeysResponse": {
      "type": "object",
      "properties": {
        "pool": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/KeysResponseKey"
          }
        }
      }
    },
    "settingsPutResponse": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        }
      }
    },
...
  }
}